requirements.txt
# Backend dependencies for Gemini AI Assistant
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
vosk>=0.3.40
gtts>=2.4.0
pydub>=0.25.1
Pillow>=10.0.0
aiofiles>=23.2.0
httpx>=0.25.0

# For audio processing
scipy>=1.11.0
numpy>=1.24.0
soundfile>=0.12.0

# For CORS and middleware
starlette>=0.27.0

# Additional utilities
typing-extensions>=4.8.0